+++ src/main/java/web/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package web.reservation.exception;
+package web.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
@@ -12,7 +12,10 @@ public enum ErrorCode {
     OUT_OF_BUSINESS_HOURS("영업 시간이 아닙니다. 영업 시간: " + BEGIN_TIME + " ~ " + LAST_TIME, HttpStatus.BAD_REQUEST),
     NOT_UNIT_OF_30_MINUTES("예약은 30분 단위로 가능합니다.", HttpStatus.BAD_REQUEST),
     RESERVATION_DUPLICATE("요청하신 시간대에 이미 예약이 있습니다.", HttpStatus.CONFLICT),
-    RESERVATION_NOT_FOUND("예약 정보를 찾을 수 없습니다.", HttpStatus.NOT_FOUND);
+    RESERVATION_NOT_FOUND("예약 정보를 찾을 수 없습니다.", HttpStatus.NOT_FOUND),
+    THEME_NOT_FOUND("테마가 존재하지 않습니다.", HttpStatus.NOT_FOUND),
+    THEME_DUPLICATE("이미 존재하는 테마입니다.", HttpStatus.CONFLICT),
+    ;
 
     private final String message;
     private final HttpStatus httpStatus;
