--- src/main/java/web/reservation/exception/ReservationExceptionAdvice.java
@@ -1,19 +1,0 @@
-package web.reservation.exception;
-
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import web.exception.ErrorResponse;
-
-@RestControllerAdvice
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class ReservationExceptionAdvice {
-
-    @ExceptionHandler(ReservationException.class)
-    public ResponseEntity<ErrorResponse> handleReservationException(ReservationException e) {
-        return ResponseEntity.status(e.getHttpStatus())
-                .body(ErrorResponse.from(e.getMessage()));
-    }
-}
