--- src/main/java/web/theme/exception/ErrorCode.java
@@ -1,20 +1,0 @@
-package web.theme.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public enum ErrorCode {
-
-    THEME_DUPLICATE("이미 존재하는 테마입니다.", HttpStatus.CONFLICT),
-    THEME_NOT_FOUND("테마가 존재하지 않습니다.", HttpStatus.NOT_FOUND),
-    ;
-
-    private final String message;
-    private final HttpStatus httpStatus;
-
-    ErrorCode(String message, HttpStatus httpStatus) {
-        this.message = message;
-        this.httpStatus = httpStatus;
-    }
-}
